выносим комментарии в отдельный модуль, привязка targetComment к генерации комментов, #4
