+++ src/view/popup.js
@@ -1,9 +1,3 @@
-import {generateFilmComments} from '../mocks/comments';
-import {getRandomInteger} from '../utils';
-import {AVAILABLE_EMOTIONS} from '../const';
-
-const COMMENTS_COUNT = getRandomInteger(1, 20);
-const comments = new Array(COMMENTS_COUNT).fill().map(generateFilmComments);
 const ACTIVE_WATCHED_STATUS_CLASS_NAME = 'film-details__watched-status--active';
 
 const generateGenresStrings = (genres) => {
@@ -12,53 +6,8 @@ const generateGenresStrings = (genres) => {
     .join('');
 };
 
-const generateComments = (filmCommentsId, comments) => {
-  return comments.filter((comment) => comment.commentId === filmCommentsId)
-    .map((comment) => createComments(comment))
-    .join('');
-};
-
-const generateEmojis = (emojis) => {
-  return emojis
-    .map((emoji) => createEmojiTemplate(emoji))
-    .join('');
-};
-
 const createGenreItemTemplate = (genre) => {
   return `<span class="film-details__genre">${genre}</span>`;
-};
-
-const createEmojiTemplate = (emoji) => {
-  return `<input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-${emoji}" value="${emoji}">
-            <label class="film-details__emoji-label" for="emoji-${emoji}">
-              <img src="./images/emoji/${emoji}.png" width="30" height="30" alt="${emoji}">
-            </label>`;
-};
-
-const createComments = (commentsCurrentData) => {
-  const {
-    emotion,
-    comment,
-    author,
-    date,
-  } = commentsCurrentData;
-  return `<li class="film-details__comment">
-            <span class="film-details__comment-emoji">
-              <img src="./images/emoji/${emotion}.png" width="55" height="55" alt="emoji-${emotion}">
-            </span>
-            <div>
-              <p class="film-details__comment-text">${comment}</p>
-              <p class="film-details__comment-info">
-                <span class="film-details__comment-author">${author}</span>
-                <span class="film-details__comment-day">${date}</span>
-                <button class="film-details__comment-delete">Delete</button>
-              </p>
-            </div>
-          </li>`;
-};
-
-const countAppropriateCommentsLength = (filmCommentsId, comments) => {
-  return comments.filter((comment) => comment.commentId === filmCommentsId).length;
 };
 
 export const createPopup = (data) => {
@@ -147,27 +96,7 @@ export const createPopup = (data) => {
         <label for="favorite" class="film-details__control-label film-details__control-label--favorite ${isFavorite ? ACTIVE_WATCHED_STATUS_CLASS_NAME : ''}">Add to favorites</label>
       </section>
     </div>
-    <div class="film-details__bottom-container">
-      <section class="film-details__comments-wrap">
-        <h3 class="film-details__comments-title">Comments
-            <span class="film-details__comments-count">${countAppropriateCommentsLength(0, comments)}</span>
-        </h3>
-        <ul class="film-details__comments-list">
-            ${generateComments(0, comments)}
-        </ul>
-        <div class="film-details__new-comment">
-          <div class="film-details__add-emoji-label">
-            <img src="images/emoji/${AVAILABLE_EMOTIONS[0]}.png" alt="${AVAILABLE_EMOTIONS[0]}" width="55" height="55">
-          </div>
-          <label class="film-details__comment-label">
-            <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-          </label>
-          <div class="film-details__emoji-list">
-            ${generateEmojis(AVAILABLE_EMOTIONS)}
-          </div>
-        </div>
-      </section>
-    </div>
+    <div class="film-details__bottom-container"></div>
   </form>
 </section>`;
 };
